requirements.txt
# https://pypi.org/project/Django/
django>=4.2.16,<4.3

# https://pypi.org/project/pillow/
pillow==11.0.0

# https://pypi.org/project/djangorestframework/
djangorestframework>=3.15.2,<3.16

# https://pypi.org/project/drf-spectacular/
drf-spectacular>=0.28.0,<0.29

# https://pypi.org/project/environs/
environs[django]>=11.2.1,<11.3

# https://pypi.org/project/django-guid/
django-guid>=3.5.0,<3.6

# https://pypi.org/project/psycopg/
psycopg==3.2.3

# https://pypi.org/project/celery/
celery[redis]>=5.4.0,<5.5

# https://pypi.org/project/django-celery-results/
django-celery-results>=2.5.1,<2.6

# https://pypi.org/project/django-extensions/
django_extensions>=3.2.3,<3.3

# https://pypi.org/project/shortuuid/
shortuuid>=1.0.13,<1.1

# https://pypi.org/project/djangorestframework-simplejwt/
djangorestframework-simplejwt>=5.3.1,<5.4

# https://pypi.org/project/filetype/
filetype>=1.2.0,<1.3